adding commit

+++ src/app/_services/token-stack.service.ts
@@ -1,3 +1,5 @@
+import { isPlatformBrowser } from '@angular/common';
+import { Inject, PLATFORM_ID } from '@angular/core';
 import { Injectable } from '@angular/core';
 import {  BehaviorSubject, Observable, of, Subject } from 'rxjs';
 import { debounceTime, delay, switchMap, tap } from 'rxjs/operators';
@@ -11,39 +13,53 @@ export class TokenStackService {
 
   private _loadingLogOut$ = new BehaviorSubject<boolean>(false);
 
-  constructor() {
+  constructor(
+    @Inject(PLATFORM_ID) private platformId,
+  ) {
    }
 
   logOut(): Observable<any> {
-    return of(1).pipe(
-      tap( () => this._loadingLogOut$.next(true)),
-      debounceTime(300),
-      tap( () => window.sessionStorage.clear()),
-      delay(1000),
-      tap( () => this._loadingLogOut$.next(false))
-    );
+    if (isPlatformBrowser(this.platformId)){
+      return of(1).pipe(
+        tap( () => this._loadingLogOut$.next(true)),
+        debounceTime(300),
+        tap( () => window.sessionStorage.clear()),
+        delay(1000),
+        tap( () => this._loadingLogOut$.next(false))
+      );
+    }
   }
 
   public saveToken( token:string ){
-    window.sessionStorage.removeItem(this.TOKEN_KEY);
-    window.sessionStorage.setItem(this.TOKEN_KEY, token);
+    if (isPlatformBrowser(this.platformId)){
+      window.sessionStorage.removeItem(this.TOKEN_KEY);
+      window.sessionStorage.setItem(this.TOKEN_KEY, token);
+    }
   }
 
   public getToken(): Observable<string> {
-    return of(sessionStorage.getItem(this.TOKEN_KEY));
+    if (isPlatformBrowser(this.platformId)){
+      return of(sessionStorage.getItem(this.TOKEN_KEY));
+    }
   }
 
   public getTokenDirect() {
-    return window.sessionStorage.getItem(this.TOKEN_KEY);
+    if (isPlatformBrowser(this.platformId)){
+      return window.sessionStorage.getItem(this.TOKEN_KEY);
+    }
   }
 
   public saveUser(user): void {
-    window.sessionStorage.removeItem(this.USER_KEY);
-    window.sessionStorage.setItem(this.USER_KEY, JSON.stringify(user));
+    if (isPlatformBrowser(this.platformId)){
+      window.sessionStorage.removeItem(this.USER_KEY);
+      window.sessionStorage.setItem(this.USER_KEY, JSON.stringify(user));
+    }
   }
 
   public getUser() {
-    return JSON.parse(sessionStorage.getItem(this.USER_KEY));
+    if (isPlatformBrowser(this.platformId)){
+      return JSON.parse(sessionStorage.getItem(this.USER_KEY));
+    }
   }
